Comments and minor Loader fixes
Added comments to all updated teleoperated files.  Changed the loader to
account for absence of a limit switch when checking whether the motor
should be stopped.  Now, the updateLoader function will rely on a
counter that tracks the number of times it has been called in order to
"time" how long the motor has been running.